use composition (with mixins) instead of inheritence to implement to_representation method and remove black fields
